feature(bulkimport,tracelistener): implement selective bulk import
This commit implements a way for CLI users to selectively bulk import
only certain modules, defaulting to the old behavior (bulk-import all)
if not specified.

Two new CLI arguments have been introduced.

`import-modules-list`, which returns a list of all the supported bulk
import modules and then exits, and `import-modules`, which accepts a
comma-separated list of modules to be imported.

This changes must be backported to the sdk-44-support support branch.